complete orders fragment create order details activity fix AddToCartBottomSheetFragment issues move log out button from profile to AccountFragment

+++ app/src/main/java/com/khedr/ecommerce/pojo/Address/AddressData.java
@@ -14,8 +14,8 @@ public class AddressData  implements Serializable {
     String region;
     String details;
     String notes;
-    int latitude = 0;
-    int longitude = 0;
+    double latitude = 0;
+    double longitude = 0;
 
     public AddressData(String name, String city, String region, String details, String notes) {
         this.name = name;
